full: fix dense representation bug
This commit fixes a bug in the "full" construction of Aho-Corasick into a
DFA. In particular, the `for_each_transition` routine was assuming that
the transitions in the "dense" (actually sparse) representation were
sorted in lexicographic order. But nothing maintained this invariant and
resulted in incorrect transitions being added to the DFA.

We fix this by forcing new transitions to be added in sorted order. It's
not clear whether this is the best strategy, but we are content with the
bug fix for now.

Fixes #37