api: test for UnwindSafe
It appears that using a trait object internally implies that the containing
type cannot be assumed to be UnwindSafe. This wasn't previously an
explicit API guarantee, but the AhoCorasick automaton in 0.6.x was
UnwindSafe and it's probably a good idea to keep that particular feature.
Here, we satisfy that by requiring Prefilter to be UnwindSafe, so that all
Prefilter trait objects are therefore also UnwindSafe.

In particular, this introduced a regression in regex that caused a regex
to not be UnwindSafe because it contained an AhoCorasick automaton.
See https://github.com/rust-lang/regex/issues/568 for details.