prefilter: add rare byte prefilter
This adds a long overdue prefilter to Aho-Corasick for rare byte
scanning. That is, in addition to the start-byte prefilter, we now have
a prefilter that tries to heuristically detect "rare" bytes in a given
set of patterns, and then use those to quickly scan to candidate
matches. It's a bit trickier than the start-byte prefilter since we need
to re-align candidate matches. For example, when searching for
`Sherlock`, the (heuristically) rarest byte is `k`, so when we find
candidate matches for `k`, we return `matchpos - 7`, since the last
occurrence of `k` in `Sherlock` occurs at offset `7`.

In order to avoid quadratic behavior, we now ensure that a prefilter
scans each byte at most once. This happens automatically for the
start-byte prefilter, but for the rare-byte prefilter, we need to keep
track of the last position that we scanned to and avoid re-running the
prefilter until Aho-Corasick makes it past that position on its own. For
this reason, prefilters now accept a `PrefilterState` directly, which
can be mutated.

Finally, the rare-byte prefilter is not always better than the
start-byte prefilter. Namely, because of the aforementioned shifting,
there is a bit more overhead, so we generally still prefer the
start-byte prefilter when we can.

Benchmarks generally remain unchanged, with the exception of these two
new benchmarks, which have massive improvements:

sherlock/dfa/name/alt6    94.56   872.7±8.38µs   650.1 MB/sec     1.00      9.2±0.07µs    60.0 GB/sec
sherlock/dfa/name/alt7    77.83   872.1±8.38µs   650.6 MB/sec     1.00     11.2±0.15µs    49.4 GB/sec

alt7, in particular, uses non-ASCII patterns. Namely, the start-byte
prefilter only works on ASCII, since the leading bytes in non-ASCII
UTF-8 encoded codepoints are typically very common, so searching on them
would lead to poor performance. But the rare-byte prefilter can pick any
byte, including comparatively rarer continuation bytes in UTF-8 encoded
codepoints.

There are undoubtedly cases where we regress performance as well, but we
are content for now given that 1) prefilters turn off if they become
ineffective and 2) we have heuristics for preferring the start-byte
prefilter over the rare-byte prefilter when both are available.