api: add anchored functionality
This commit adds a new option to the AhoCorasick API for anchoring
searches. When enabled, this efficiently implements the requirement that
all matches must start at the beginning of the haystack.

The primary way we achieve this is by simply eliding all failure
transitions, and setting them instead to the dead state. Effectively,
this turns the Aho-Corasick automaton into a traditional trie. Once
searching sees a dead state, it quits. (This required a few changes,
since we previously asserted that dead states could only follow match
states, which is now no longer true.)